requirements_dev.txt
pip==9.0.1
bumpversion==0.5.3
wheel==0.30.0
watchdog==0.8.3
flake8==3.5.0
tox==2.9.1
coverage==4.5.1
Sphinx==1.7.1
twine==1.10.0

pytest==3.4.2
pytest-runner==2.11.1
autopep8==1.3.5
responses==0.9.0
ipython==6.4.0
apache-airflow==1.9.0
SQLAlchemy==1.2.8
rows==0.3.1
psycopg2-binary==2.7.4
